+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 # Current version
 
+* Fix changing email
+* added news author in emails
 * Fixed events' start time in iCalc
 
 # Version 18.01.05